adds new e2e test with wdio
