+++ scripts/e2e-v6.sh
@@ -2,13 +2,13 @@
 
 source ${OKTA_HOME}/${REPO}/scripts/setup.sh
 
-setup_service xvfb start
 setup_service java 1.8.222
-setup_service google-chrome-stable 87.0.4280.66-1
+setup_service google-chrome-stable 89.0.4389.72-1
 
 export TEST_SUITE_TYPE="junit"
 export TEST_RESULT_FILE_DIR="${REPO}/test-reports/e2e"
 
+export CI=true
 export ISSUER=https://samples-javascript.okta.com/oauth2/default
 export SPA_CLIENT_ID=0oapmwm72082GXal14x6
 export USERNAME=george@acme.com
@@ -19,10 +19,7 @@ if ! yarn add -DW --ignore-scripts @okta/okta-auth-js@^6; then
   exit ${FAILED_SETUP}
 fi
 
-sh ~/okta/okta-vue/scripts/update_se_drivers.sh
-cd ~/okta/okta-vue/test/e2e
-
-if ! yarn e2e; then
+if ! yarn test:e2e; then
   echo "e2e tests failed! Exiting..."
   exit ${TEST_FAILURE}
 fi
