--- test/app/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@okta/test.app",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "start": "vue-cli-service serve",
-    "build": "vue-cli-service build"
-  },
-  "dependencies": {
-    "@okta/okta-vue": "*",
-    "vue": "^3.2.31",
-    "vue-router": "^4.0.14"
-  },
-  "devDependencies": {
-    "@vue/cli-plugin-router": "~5.0.8",
-    "@vue/cli-plugin-typescript": "~5.0.8",
-    "@vue/cli-service": "~5.0.8",
-    "@vue/compiler-sfc": "^3.2.31",
-    "source-map-loader": "^1.1.3"
-  }
-}
