--- test/app/src/components/Home.vue
@@ -1,12 +1,0 @@
-<template>
-  <div class="home">
-  </div>
-</template>
-  
-<script lang="ts">
-import { defineComponent } from 'vue'
-
-export default defineComponent({
-  name: 'Home'
-})
-</script>
