--- test/app/vue.config.js
@@ -1,40 +1,0 @@
-require('@okta/env').setEnvironmentVarsFromTestEnv();
-
-process.env.CLIENT_ID = process.env.SPA_CLIENT_ID || process.env.CLIENT_ID;
-
-const env = {};
-
-// List of environment variables made available to the app
-[
-  'ISSUER',
-  'CLIENT_ID',
-  'OKTA_TESTING_DISABLEHTTPSCHECK',
-  'USE_INTERACTION_CODE',
-].forEach((key) => {
-  if (!process.env[key]) {
-    throw new Error(`Environment variable ${key} must be set. See README.md`);
-  }
-  env[key] = process.env[key];
-});
-
-module.exports = {
-  configureWebpack: {
-    module: {
-      rules: [
-        {
-          test: /\.js$/,
-          use: ['source-map-loader'],
-          enforce: 'pre'
-        },
-      ]
-    }
-  },
-  chainWebpack: config => {
-    config.plugin('define').tap(definitions => {
-      definitions[0] = Object.assign(definitions[0], {
-        CONFIG: JSON.stringify(env)
-      })
-      return definitions
-    })
-  }
-}
