--- test/e2e-wdio/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "@okta/test.e2e-wdio",
-  "version": "1.0.0",
-  "private": true,
-  "scripts": {
-    "test": "node runner.js",
-    "wdio": "wdio run wdio.conf.js",
-    "test:harness": "SAMPLE_NAME=harness yarn test"
-  },
-  "devDependencies": {
-    "@babel/cli": "^7.19.3",
-    "@babel/core": "^7.19.3",
-    "@babel/preset-env": "^7.19.3",
-    "@babel/register": "^7.18.9",
-    "@wdio/cli": "^7.25.1",
-    "@wdio/jasmine-framework": "^7.25.1",
-    "@wdio/junit-reporter": "^7.25.1",
-    "@wdio/local-runner": "^7.25.1",
-    "@wdio/selenium-standalone-service": "^7.25.1",
-    "@wdio/spec-reporter": "^7.25.1",
-    "cross-spawn-with-kill": "^1.0.0",
-    "eslint-plugin-jasmine": "^4.1.1",
-    "url": "^0.11.0",
-    "wait-on": "^6.0.1"
-  }
-}
