--- test/e2e/.gitignore
@@ -1,2 +1,0 @@
-reports
-selenium-server.log*
