+++ test/e2e/page-objects/test-harness-app/protected-page.js
